pyproject.toml
[project]
name = "aviation-stack-mcp"
version = "0.1.0"
description = "This is a mcp server for aviationstack flight APIs"
authors = [
    {name = "Mohit Jain",email = "207972266+mohitatust@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp",
    "rich"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ruff = "^0.11.12"
black = "^25.1.0"

